pyproject.toml
# AGE-YOLO 🚀: Accurate, Generalized, and Efficient Object Detector
# Optimized for Steel Surface Defect Detection

[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "age_yolo"
version = "1.0.0"
description = "AGE-YOLO: Multi-Axis Network with Adaptive Context Fusion for Steel Surface Defect Detection"
readme = "README.md"
requires-python = ">=3.8"
license = { "text" = "AGPL-3.0" }
keywords = [
    "Steel-Surface-Defect-Detection",
    "GC10-DET",
    "FC-MANet",
    "ACF-Module",
    "SIFHead"
]
authors = [
    { name = "gnmtdt" },
]
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.23.0",
    "matplotlib>=3.3.0",
    "opencv-python>=4.6.0",
    "pillow>=7.1.2",
    "pyyaml>=5.3.1",
    "torch>=1.8.0",
    "torchvision>=0.9.0",
    "tqdm>=4.64.0",
    "psutil",
    "pandas>=1.1.4",
    "seaborn>=0.11.0",
]

[project.urls]
"Source" = "https://github.com/gnmtdt/AGE-YOLO"

[tool.setuptools]
packages = { find = { where = ["."], include = ["ultralytics", "ultralytics.*"] } }